--- components/shared/tweet/tweet.module.css
@@ -1,9 +1,0 @@
-.card {
-  padding: 1.5rem;
-  padding-bottom: 1rem;
-  border-radius: 0.5rem;
-  border-width: 1px;
-  border-color: #d1d5db;
-  backdrop-filter: blur(8px);
-  /* shadow-[inset_10px_-50px_94px_0_rgb(199,199,199,0.2)] */
-}
