+++ lib/twitter/index.ts
@@ -11,5 +11,4 @@ export const homepageTweets = [
   "1580891783277395970",
   "1582956754425421825",
   "1574164441498894339",
-  //   "1580479816071270401",
 ];
