+++ package.json
@@ -53,6 +53,7 @@
     "react-dom": "18.1.0",
     "react-dom-confetti": "^0.2.0",
     "react-hot-toast": "^2.4.0",
+    "react-parallax-tilt": "^1.7.68",
     "react-spring": "^9.5.4",
     "react-textarea-autosize": "^8.3.4",
     "stripe": "^10.12.0",
